groq-vision-model/requirements.txt
fastapi
groq
python-dotenv
uvicorn
pydantic
requests
